move testMetricsChange to MetricST and add check that enabling metric triggers rolling update
